move snapshot tests into test_integration
