+++ tox.ini
@@ -23,7 +23,7 @@ envlist =
     {py27,py35,py36,py37,py38}-tracer
     {py27,py35,py36,py37,py38,py39}-profile{,-gevent}
     {py27,py35,py36,py37,py38,py39}-profile-minreqs{,-gevent}
-    {py27,py35,py36,py37,py38}-integration{,-v5}
+    {py27,py35,py36,py37,py38}-integration{,-v5,snapshot}
     py{27,35,36,37,38}-vendor
     {py27,py35,py36,py37,py38}-ddtracerun
     {py27,py35,py36,py37,py38}-test_logging
@@ -166,7 +166,9 @@ usedevelop = true
 
 setenv =
     profile-gevent: DD_PROFILE_TEST_GEVENT=1
+    integration: AGENT_VERSION=latest
     integration-v5: AGENT_VERSION=5
+    integration-snapshot: AGENT_VERSION=testagent
     # DEV: envs to test migration of django instrumentation to ddtrace config api
     django_contrib_migration: TEST_DATADOG_DJANGO_MIGRATION=1
     bottle_contrib_autopatch: DATADOG_SERVICE_NAME=bottle-app
